component-ize day hero day linear chart
