+++ components/styles/DataJournal.module.css
@@ -4,12 +4,12 @@
   align-items: center;
 }
 
-.linearContainer {
+.sliceContainer {
   display: flex;
   width: 100%;
   height: 100px;
 }
 
-.linearDay {
+.daySlice {
   flex-grow: 1;
 }
