+++ pages/datajournal/index.tsx
@@ -3,20 +3,9 @@ import {usePapaParse} from 'react-papaparse';
 import moment from 'moment';
 import * as React from 'react';
 
-import styles from '../../components/styles/HowImDoing.module.css';
+import {Day, DaySlice} from '../../components/datajournal/DayTile';
 
-
-interface Day {
-  date: moment.Moment;
-  rating: number;
-  create: boolean;
-  care: boolean;
-  talk: boolean;
-  move: boolean;
-  work: boolean;
-  read: boolean;
-  write: boolean;
-};
+import styles from '../../components/styles/DataJournal.module.css';
 
 
 const HowImDoing: NextPage = () => {
@@ -44,30 +33,13 @@ const HowImDoing: NextPage = () => {
 
   React.useEffect(loadDayDataFromCSV, []);
 
-  const ratingToColorMap = [
-    "#999999",
-    "#000000",
-    "#FF0000",
-    "#FFA500",
-    "#00FF00",
-    "#1E90FF"
-  ];
-
   return <>
-    <div className={styles.linearContainer}>
-      {days.map(day => {
-        const dateLabel = day.date.format("MM/DD/YYYY");
-        return <div
-          key={dateLabel}
-          title={dateLabel}
-          className={styles.linearDay}
-          style={{backgroundColor: ratingToColorMap[day.rating]}}>
-        </div>
-      })}
+    <div className={styles.sliceContainer}>
+      {days.map(day => <DaySlice day={day} />)}
     </div>
     <header className={styles.headerContainer}>
       <h1>So...How Have You Been?</h1>
-      <h2>a makeshift journal</h2>
+      <h2>a makeshift journal of {days.length} days and counting</h2>
     </header>
   </>;
 }
